requirements.txt
django==5.1.2
django-ninja==1.3.0
django-ninja-jwt==5.3.4
python-dotenv==1.0.1
coverage==7.6.4
ruff==0.7.1

vertexai==1.70.0